Remove `Runner`, fold implementation into `CmdRunner`
